src/main/java/com/example/api/service/AccountService.java
package com.example.api.service;

import java.util.Date;
import com.example.api.domain.Account;
import com.example.api.repository.AccountRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.stereotype.Service;
import org.springframework.web.server.ResponseStatusException;

@Service
public class AccountService {

    @Autowired
    private AccountRepository accountRepository;

    public Account registerAccount(Account account) {
        if (accountRepository.existsByEmail(account.getEmail())) {
            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "사용중인 이메일입니다.");
        }

        if (account.getPassword().length() < 8) {
            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "비밀번호는 8자리 이상이어야 합니다.");
        }

        account.setRegisteredDatetime(new Date());

        return accountRepository.save(account);
    }
}

